[MRESOLVER-231] Extend smart checksum support
Refactor and make it pluggable. Also, make it work with Central.

This change broadens "smart checksum" support in Resolver, and
gets them from HTTP response header, if present. So far only
Nexus2 was supported, but this now adds more implementations,
so for example Central is supported now as well, hence, to
download a JAR from Central will now take only 1 HTTP request
and not 2.